split WFS into small files in `wfs` directory
