+++ soapy/wfs/__init__.py
@@ -1,0 +1,2 @@
+from . import wfs
+WFS = wfs # Compatability
